move tantan's stuff into its own module.
